requirements.txt
flask
sockets
flask-socketio
pyopenssl
pyjwt